+++ nextjs-graphql-hooks.php
@@ -13,21 +13,8 @@
  * Tested up to: 6.4
  * Network: false
  * Requires Plugins: wp-graphql, elementor
- * License: GPL v2 or later
- * License URI: https://www.gnu.org/licenses/gpl-2.0.html
- *
- * NextJS GraphQL Hooks is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * any later version.
- *
- * NextJS GraphQL Hooks is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with NextJS GraphQL Hooks. If not, see https://www.gnu.org/licenses/gpl-2.0.html.
+ * License: Polyform Noncommercial License 1.0.0
+ * License URI: https://polyformproject.org/licenses/noncommercial/1.0.0/
  *
  * @package NextJSGraphQLHooks
  * @since 1.0.0