+++ includes/Updater.php
@@ -8,7 +8,7 @@
  * @package NextJSGraphQLHooks
  * @since 1.0.0
  * @author Silver Assist
- * @version 1.0.3
+ * @version 1.0.4
  * @license Polyform Noncommercial License 1.0.0
  */